Fix the bug that can't send mail to remind users of beans.
