+++ src/httprest.py
@@ -559,6 +559,20 @@ def listphynodes_monitor(cur_user, user, form):
     res['allnodes'] = G_nodemgr.get_allnodes()
     return json.dumps({'success':'true', 'monitor':res})
 
+@app.route("/beans/mail/", methods=['POST'])
+@worker_ip_required
+def beans_mail():
+    logger.info("handle request: beans/mail/")
+    addr = request.form.get("to_address",None)
+    username = request.form.get("username",None)
+    beans = request.form.get("beans",None)
+    if addr is None or username is None or beans is None:
+        return json.dumps({'success':'false', 'message':"to_address,username and beans fields are required!"})
+    else:
+        logger.info("send email to "+addr+" and username:"+username+" beans:"+beans)
+        beansapplicationmgr.send_beans_email(addr,username,int(beans))
+        return json.dumps({'success':'true'})
+
 @app.route("/beans/<issue>/", methods=['POST'])
 @login_required
 def beans_apply(cur_user,user,form,issue):
